otel: add build tooling to include otel code
Adds the --otel flag to the configure command and the various build time
variables and checks that are needed in this flow.

It also includes the nxt_otel.c and nxt_otel.h files that are needed for
the rest of Unit to talk to the compiled static library that's generated
from the rust crate.

Signed-off-by: Gabor Javorszky <g.javorszky@f5.com>
Signed-off-by: Ava Hahn <a.hahn@f5.com>

Co-authored-by: Ava Hahn <a.hahn@f5.com>