tools: changing clang-format to ideal per tool
The tool in question is https://github.com/javorszky/clang-format-finder-unit,
which tried most combination of values for most options and came to an
ideal format file that changes the least amount of lines. See the other
repository to see how that works.

This doesn't mean that this format file is in alignment with what we
want, it means this is the least invasive current setting.