Packages: support rpm 4.20+
In 4.20, rpm changed the way build roots are named.  Our packaging
relies on the symlinks being created as a part of a rule to make sure
targets are being built correctly.

This patch implements (a hacky) way to check what the build root should
be named on a currently running rpm version and adjusts the symlinks
accordingly.