Fix GCC 15 Wunterminated-string-initialization issues
When building with the upcoming GCC 15, we are hit by the new
Wunterminated-string-initialization compiler warning (however, due to
-Werror, this is a build failure). E.g.

  src/nxt_string.c: In function ‘nxt_encode_uri’:
  src/nxt_string.c:601:36: error: initializer-string for array of ‘unsigned char’ is too long [-Werror=unterminated-string-initialization]
    601 |     static const u_char  hex[16] = "0123456789ABCDEF";
        |                                    ^~~~~~~~~~~~~~~~~~

We have created a character array containing 16 elements, but is not NUL
terminated. While this type of (lookup/translation table) use case is
perfectly valid and we could simply disable the warning, it is probably
still a useful warning to have enabled.

There is a 'nonstring' attribute, but that doesn't influence this
warning. Well not yet at least... but that looks to change!

Lets just do what nginx does and create them NUL terminated. Except in
src/nxt_http_parse.c where they *really* want to be non-NUL
terminated and we use *arrays* of chars.

Link: <https://gcc.gnu.org/onlinedocs/gcc-14.2.0/gcc/Common-Variable-Attributes.html>
Link: <https://gcc.gnu.org/bugzilla/show_bug.cgi?id=117178#c21>
Signed-off-by: Andrew Clayton <a.clayton@nginx.com>