python: Fix Litestar WebSockets compatibility
It was reported on GitHub that Unit was unable to work with WebSockets
under Litestar Python applications.

This was due to Unit sending a 'method' variable in the WebSocket's
connection scope, which Litestar was interpreting as being a normal HTTP
connection.

The ASGI WebSocket specification makes no mention about setting a
'method', so let's not send it on WebSockets.

Also tested this change with basic ASGI WebSockets and FastAPI
WebSockets and obviously pytests still pass.

Closes: https://github.com/nginx/unit/issues/1507
Link: <https://asgi.readthedocs.io/en/latest/specs/www.html#websocket-connection-scope>
Signed-off-by: Andrew Clayton <a.clayton@nginx.com>