http: Fix WebSockets with Firefox
Firefox (134 at least) was unable to open a WebSocket connection to Unit
due to it sending a Connection header of

  Connection: keep-alive, Upgrade

However in Unit we were expecting only a single value in the header.

Fix the 'Connection' parsing in nxt_h1p_connection() to address this.

We keep the check for 'close' basically the same as that really should
be the only value.

Then we check for both of 'keep-alive' and 'upgrade'.

Closes: https://github.com/nginx/unit/issues/772
Signed-off-by: Andrew Clayton <a.clayton@nginx.com>