fix: Notify app of client disconnection when request is in progress.
Previously, the app was not notified when the client disconnected. This caused
issues especially in cases of websocket connections and SSE Events where the app
continued to send data to the router, which could not deliver it to the
client due to the disconnection.

Changes made:
   1. Added functionality to send a port message to notify the app of client
      disconnection.
   2. For now handled and tested this scenario only for ASGI websockets and
      ASGI HTTP, ensuring that if the app detects a client disconnection, it follows
      the ASGI specification.

This ensures that the app is properly informed of client disconnections
and can handle them according to the ASGI spec.