node.js: Fixed issue with duplicate headers in response
It fixes losing context in response in cases when there are 2 or more
headers with the same name.  The prev implementation used to use foreach
function which uses local lexical environment and did not find
this.headers_len locally, which causes crash of the http server module.

It was replaced with a for loop in order to make access for this.headers_len
variable and improve performance of calculation.

Closes: https://github.com/nginx/unit/issues/1621
Signed-off-by: Andrew Clayton <a.clayton@nginx.com>