wasm-wc: Update to wasmtime 35.0.0
This is mainly just to be on the latest version for the next release of
Unit.

This required some changes to the language module as described here
<https://github.com/bytecodealliance/wasmtime/pull/10016>.

We also add unnecessary_transmutes to the list of allowed linters to
quell warnings like

  warning: unnecessary transmute
     --> /home/andrew/src/unit/src/wasm-wasi-component/target/debug/build/wasm-wasi-component-9ae3c2c94201e6be/out/bindings.rs:440:13
      |
  440 | ...   ::std::mem::transmute(self._bitfield_1.get(0usize, 24u8) as u32)
      |       ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ help: replace this with: `u32::cast_signed(self._bitfield_1.get(0usize, 24u8) as u32)`
      |
      = note: `#[warn(unnecessary_transmutes)]` on by default

Also, because this is new in rustc 1.88, to avoid warnings on older
compilers the simplest thing is to just add unknown_lints to the list.

Link: <https://lists.gnu.org/archive/html/qemu-rust/2025-07/msg00006.html>
Signed-off-by: Andrew Clayton <a.clayton@nginx.com>