rust, wasm-wc, otel: Update the slab crate
This updates the slab crate to 0.4.11. This is the same as commit
d413a255 ("tools/unitctl: Bump the slab crate from 0.4.10 to 0.4.11")
which updated the slab crate in unitctl.

That was done by dependabot. It turns out this update is also needed for
OTEL and wasm-wasi-component.

Unfortunately it seems dependabot is b0rked and hasn't created
pull-requests to update those.

Signed-off-by: Andrew Clayton <a.clayton@nginx.com>